pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "yfinance>=0.2.55",
]
